feat: implement relative date formatting for last indexed timestamps in connector views
